refactor(test): replace fixture with minimal flat cycle/ package
Replace the multi-module app/ + zoo-sdk-stub/ fixture with a flat
cycle/ package (5 files, single JavaParserTypeSolver root).

The 3-node import structure is key: ZooTestConstants imports
ZooQueryConstants.* first, then ElephantTestConstants.*. When
ZooQueryConstants.* returns unsolved it calls resolvedStaticImport.remove(),
flushing the guard set mid-search. This lets ZooTestConstants re-enter
the search, causing unbounded recursion — the exact path our
RESOLVING_MEMBER_IN_TYPE_ONLY fix breaks.

Verified: assertDoesNotThrow FAILS on test-cyclic-static-import-expect-stackoverflow
          (no fix) and PASSES on this branch (fix applied).
