+++ javaparser-symbol-solver-testing/src/test/resources/static_import_cycle_fixture/ElephantBuilder.java
@@ -1,0 +1,11 @@
+import static ZooTestConstants.*;
+
+import java.util.Objects;
+
+public class ElephantBuilder {
+
+    public ElephantBuilder withLionObjCode() {
+        Objects.requireNonNull(Lion.OBJCODE);
+        return this;
+    }
+}
