+++ javaparser-symbol-solver-testing/src/test/resources/static_import_cycle_fixture/ElephantTestConstants.java
@@ -1,0 +1,7 @@
+import static ZooTestConstants.*;
+
+@SuppressWarnings({"unused", "WeakerAccess"})
+public final class ElephantTestConstants {
+    public static final String ELEPHANT = "elephant";
+    public static final String ELEPHANT_ID = ELEPHANT + ZooQueryConstants.ID;
+}
