+++ javaparser-symbol-solver-testing/src/test/resources/static_import_cycle_fixture/Lion.java
@@ -1,5 +1,3 @@
-package com.example.zoo.sdk.internal;
-
 public class Lion {
     public static final String OBJCODE = "lion";
 }
