+++ javaparser-symbol-solver-testing/src/test/resources/static_import_cycle_fixture/ZooQueryConstants.java
@@ -1,5 +1,3 @@
-package com.example.zoo.common;
-
 public interface ZooQueryConstants {
     String ID = "ID";
     String NAME = "name";
