+++ javaparser-symbol-solver-testing/src/test/resources/static_import_cycle_fixture/ZooTestConstants.java
@@ -1,0 +1,4 @@
+import static ZooQueryConstants.*;
+import static ElephantTestConstants.*;
+
+public final class ZooTestConstants {}
