--- javaparser-symbol-solver-testing/src/test/resources/static_import_cycle_fixture/app/src/test/java/junit4/zoo/builders/Elephant.java
@@ -1,7 +1,0 @@
-package junit4.zoo.builders;
-
-public class Elephant {
-    public void setTaskID(String taskID) {}
-    public void setDocObjCode(String docObjCode) {}
-    public void setObjID(String objID) {}
-}
