--- javaparser-symbol-solver-testing/src/test/resources/static_import_cycle_fixture/app/src/test/java/junit4/zoo/builders/ElephantBuilder.java
@@ -1,18 +1,0 @@
-package junit4.zoo.builders;
-
-import com.example.zoo.sdk.internal.*;
-
-import static junit4.zoo.constants.ZooTestConstants.*;
-
-@SuppressWarnings("unused")
-public class ElephantBuilder {
-
-    private Elephant apiObject = new Elephant();
-
-    public ElephantBuilder lionID(String lionID) {
-        apiObject.setTaskID(lionID);
-        apiObject.setDocObjCode(Lion.OBJCODE);
-        apiObject.setObjID(lionID);
-        return this;
-    }
-}
