--- javaparser-symbol-solver-testing/src/test/resources/static_import_cycle_fixture/app/src/test/java/junit4/zoo/constants/ZooRegexConstants.java
@@ -1,13 +1,0 @@
-package junit4.zoo.constants;
-
-@SuppressWarnings({"unused", "WeakerAccess"})
-public final class ZooRegexConstants {
-    public static final String DOT = ".";
-    public static final String COLON = ":";
-    public static final String UNDERSCORE = "_";
-    public static final String SPACE = " ";
-    public static final String FORWARD_SLASH = "/";
-
-    private ZooRegexConstants() {
-    }
-}
