--- javaparser-symbol-solver-testing/src/test/resources/static_import_cycle_fixture/app/src/test/java/junit4/zoo/constants/ZooTestConstants.java
@@ -1,13 +1,0 @@
-package junit4.zoo.constants;
-
-import static com.example.zoo.common.ZooQueryConstants.*;
-import static junit4.zoo.constants.ZooRegexConstants.*;
-import static junit4.zoo.constants.elephant.ElephantTestConstants.*;
-
-@SuppressWarnings("WeakerAccess")
-public final class ZooTestConstants {
-    public static final String ID = "ID";
-    public static final String UNDERSCORE = "_";
-    public static final String PARENT_ID = "parentID";
-    public static final String TEST = "test";
-}
