--- javaparser-symbol-solver-testing/src/test/resources/static_import_cycle_fixture/app/src/test/java/junit4/zoo/constants/elephant/ElephantTestConstants.java
@@ -1,11 +1,0 @@
-package junit4.zoo.constants.elephant;
-
-import static com.example.zoo.common.ZooQueryConstants.*;
-import static junit4.zoo.constants.ZooRegexConstants.*;
-import static junit4.zoo.constants.ZooTestConstants.*;
-
-@SuppressWarnings({"unused", "WeakerAccess"})
-public final class ElephantTestConstants {
-    public static final String ELEPHANT = "elephant";
-    public static final String ELEPHANT_ID = ELEPHANT + com.example.zoo.common.ZooQueryConstants.ID;
-}
